requirements.txt
pyairtable==3.1.1
st-supabase-connection==2.0.0
st-copy-to-clipboard==0.1.6
linkup-sdk==0.2.4
markdown==3.7
pydantic==2.6.0
boto3==1.34.144
nltk==3.8.1
scikit-learn==1.2.2
openai==1.77.0
anthropic==0.50.0
pandas==2.2.0
pydantic==2.6
openpyxl==3.1.5
tenacity==8.4.2
tqdm>4
langchain==0.2.11
langchain-community==0.2.10
faiss-cpu>=1.7
sentence-transformers==3.0.1
rank-bm25==0.2.2
httpx==0.27.2
hf-xet==1.1.0
Flask==3.0.3
gunicorn==22.0.0
Flask-WTF==1.2.1
WTForms==3.1.2
python-dotenv==1.0.1
